Fixed overflow in GetPerformanceCounter()
Summary: I ran UBSAN on the tests and this was one of the easier fixes. Bitcoin compiles as a 32-bit app and as such this usage ends up with an overflow. Note that this bug isn't super-critical as the GetPerformanceCounter() function is only used to seed random number generators, but it still is a bug nonetheless.

Test Plan: make check and rpc-tests.py

Reviewers: freetrader, deadalnix, #bitcoin_abc

Reviewed By: deadalnix, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D472

+++ src/random.cpp
@@ -36,7 +36,7 @@ static inline int64_t GetPerformanceCounter() {
 #else
     timeval t;
     gettimeofday(&t, nullptr);
-    nCounter = (int64_t)(t.tv_sec * 1000000 + t.tv_usec);
+    nCounter = t.tv_sec * 1000000LL + t.tv_usec;
 #endif
     return nCounter;
 }
